requirements.txt
asgiref==3.2.3
cffi==1.13.2
cryptography==2.8
Django==3.0.2
django-cors-headers==3.2.1
django-environ==0.4.5
django-rest-knox==4.1.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djoser==2.0.3
mysqlclient==1.4.6
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.3
python-dotenv==0.10.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
